refactor(ui): canonicalize totp setup module filenames
